[Bind] Move bind2 -> bind
