+++ tests/gold/test_bind_basic_split_verilog.mlir.tpl
@@ -1,7 +1,7 @@
 module attributes {circt.loweringOptions = "locationInfoStyle=none,omitVersionComment"} {
-    hw.module @TopBasicAsserts_mlir(in %I: i1, in %O: i1, in %other: i1) attributes {output_file = #hw.output_file<"$cwd/build/TopBasicAsserts_mlir.v">, output_filelist = #hw.output_filelist<"$cwd/build/test_bind2_basic_split_verilog_bind_files.list">} {
+    hw.module @TopBasicAsserts_mlir(in %I: i1, in %O: i1, in %other: i1) attributes {output_file = #hw.output_file<"$cwd/build/TopBasicAsserts_mlir.v">, output_filelist = #hw.output_filelist<"$cwd/build/test_bind_basic_split_verilog_bind_files.list">} {
     }
-    hw.module @Top(in %I: i1, out O: i1) attributes {output_file = #hw.output_file<"$cwd/build/test_bind2_basic_split_verilog.v">} {
+    hw.module @Top(in %I: i1, out O: i1) attributes {output_file = #hw.output_file<"$cwd/build/test_bind_basic_split_verilog.v">} {
         %1 = hw.constant -1 : i1
         %0 = comb.xor %1, %I : i1
         hw.instance "TopBasicAsserts_mlir_inst0" sym @Top.TopBasicAsserts_mlir_inst0 @TopBasicAsserts_mlir(I: %I: i1, O: %0: i1, other: %I: i1) -> () {doNotPrint = true}
